html_clean {rmdformats}	R Documentation

Convert to an HTML document

Description

Format for converting from R Markdown to an HTML document.

Usage

html_clean(
  fig_width = 6,
  fig_height = 6,
  fig_caption = TRUE,
  highlight = "kate",
  lightbox = TRUE,
  thumbnails = TRUE,
  gallery = FALSE,
  pandoc_args = NULL,
  md_extensions = NULL,
  toc_depth = 2,
  toc = TRUE,
  mathjax = "rmdformats",
  use_bookdown = FALSE,
  ...
)

Arguments

fig_width
Default width (in inches) for figures

fig_height
Default width (in inches) for figures

fig_caption
TRUE to render figures with captions

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "textmate". Pass NULL to prevent syntax highlighting.

lightbox
if TRUE, add lightbox effect to content images

thumbnails
if TRUE display content images as thumbnails

gallery
if TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display

pandoc_args
arguments passed to the pandoc_args argument of rmarkdown html_document

md_extensions
arguments passed to the md_extensions argument of rmarkdown html_document

toc_depth
adjust table of contents depth

toc
if TRUE, display a table of contents

mathjax
set to NULL to disable Mathjax insertion

use_bookdown
if TRUE, uses html_document2 instead of html_document, thus providing numbered sections and cross references

...
Additional function arguments passed to R Markdown html_document

Details

Styling and features are very similar to the ones from the great knitrBootstrap package by Jim Hester : https://github.com/jimhester/knitrBootstrap

Value

R Markdown output format to pass to render

[Package rmdformats version 0.3.7 Index]