Material design - bootstrap HTML output format

Format for converting from R Markdown to an HTML document with a material design theme.

material(
  fig_width = 6,
  fig_height = 6,
  fig_caption = TRUE,
  highlight = "kate",
  lightbox = TRUE,
  thumbnails = TRUE,
  gallery = FALSE,
  cards = TRUE,
  use_bookdown = FALSE,
  pandoc_args = NULL,
  md_extensions = NULL,
  mathjax = "rmdformats",
  ...
)

Arguments

fig_width

Default width (in inches) for figures

fig_height

Default width (in inches) for figures

fig_caption

TRUE to render figures with captions

highlight

Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "textmate". Pass NULL to prevent syntax highlighting.

lightbox

if TRUE, add lightbox effect to content images

thumbnails

if TRUE display content images as thumbnails

gallery

if TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display

cards

if TRUE, sections will be presented as distinct and animated cards

use_bookdown

if TRUE, uses html_document2 instead of html_document, thus providing numbered sections and cross references

pandoc_args

arguments passed to the pandoc_args argument of rmarkdown html_document

md_extensions

arguments passed to the md_extensions argument of rmarkdown html_document

mathjax

set to NULL to disable Mathjax insertion

...

Additional function arguments passed to R Markdown html_document

Value

R Markdown output format to pass to render

Details

JavaScript and CSS taken and adapted from the Material design theme for Bootstrap 3 project : https://github.com/FezVrasta/bootstrap-material-design.