pilltabs {rmdformats}	R Documentation

Given a cross-table, outputs HTML code to display several views of with a tabbed interface

Description

Given a two dimensions contingency tables, this function outputs HTML code to display, within a dynamic tabbed interface, the count, line row percentages, column percentages and chi-squared residuals tables.

Usage

pilltabs(
  tab,
  count = TRUE,
  rows = TRUE,
  cols = TRUE,
  chisq = TRUE,
  resid = TRUE,
  row.names = TRUE
)

Arguments

tab
a two dimensions table object

count
wether or not to the display the count table

rows
wether or not to the display the row percentages table

cols
wether or not to the display the column percentages table

chisq
wether or not to the display the table chi-squared test results

resid
wether or not to the display the chi-squared residuals table

row.names
wether or not to display the table row names

Details

The function is intended to be called inside an rmarkdown document.

Value

No value is returned.

Author(s)

Julien Barnier <julien.barnier@ens-lyon.fr>

Examples

data(airquality)
tab <- table(airquality$Month, airquality$Ozone > 25)
pilltabs(tab)

[Package rmdformats version 0.3.7 Index]